Start porting over shipyard to MyBot
