+++ model.hpp
@@ -8,7 +8,7 @@
 struct ActorCriticNetwork : torch::nn::Module {
 public:
 
-    ActorCriticNetwork()
+    ActorCriticNetwork(bool training)
     :   conv1(torch::nn::Conv2dOptions(NUMBER_OF_FRAMES, 32, /*kernel_size=*/7)),
         conv2(torch::nn::Conv2dOptions(32, 64, /*kernel_size=*/3)),
         conv3(torch::nn::Conv2dOptions(64, 64, /*kernel_size=*/3)),
@@ -36,40 +36,42 @@ public:
         device = torch::Device(device_type);
         this->to(device);
 
-        //Print out network information at beginning of run
-        std::cout << "Conv1: (";
-        std::cout << conv1.get()->weight.size(0) << ", ";
-        std::cout << conv1.get()->weight.size(1) << ", ";
-        std::cout << conv1.get()->weight.size(2) << ", ";
-        std::cout << conv1.get()->weight.size(3) << ")" << std::endl;
+        if(training) {
+            //Print out network information at beginning of run
+            std::cout << "Conv1: (";
+            std::cout << conv1.get()->weight.size(0) << ", ";
+            std::cout << conv1.get()->weight.size(1) << ", ";
+            std::cout << conv1.get()->weight.size(2) << ", ";
+            std::cout << conv1.get()->weight.size(3) << ")" << std::endl;
 
-        std::cout << "Conv2: (";
-        std::cout << conv2.get()->weight.size(0) << ", ";
-        std::cout << conv2.get()->weight.size(1) << ", ";
-        std::cout << conv2.get()->weight.size(2) << ", ";
-        std::cout << conv2.get()->weight.size(3) << ")" << std::endl;
+            std::cout << "Conv2: (";
+            std::cout << conv2.get()->weight.size(0) << ", ";
+            std::cout << conv2.get()->weight.size(1) << ", ";
+            std::cout << conv2.get()->weight.size(2) << ", ";
+            std::cout << conv2.get()->weight.size(3) << ")" << std::endl;
 
-        std::cout << "Conv3: (";
-        std::cout << conv3.get()->weight.size(0) << ", ";
-        std::cout << conv3.get()->weight.size(1) << ", ";
-        std::cout << conv3.get()->weight.size(2) << ", ";
-        std::cout << conv3.get()->weight.size(3) << ")" << std::endl;
+            std::cout << "Conv3: (";
+            std::cout << conv3.get()->weight.size(0) << ", ";
+            std::cout << conv3.get()->weight.size(1) << ", ";
+            std::cout << conv3.get()->weight.size(2) << ", ";
+            std::cout << conv3.get()->weight.size(3) << ")" << std::endl;
 
-        std::cout << "fc1: (";
-        std::cout << fc1.get()->weight.size(0) << ", ";
-        std::cout << fc1.get()->weight.size(1) << ")" << std::endl;
+            std::cout << "fc1: (";
+            std::cout << fc1.get()->weight.size(0) << ", ";
+            std::cout << fc1.get()->weight.size(1) << ")" << std::endl;
 
-        std::cout << "fc2: (";
-        std::cout << fc2.get()->weight.size(0) << ", ";
-        std::cout << fc2.get()->weight.size(1) << ")" << std::endl;
+            std::cout << "fc2: (";
+            std::cout << fc2.get()->weight.size(0) << ", ";
+            std::cout << fc2.get()->weight.size(1) << ")" << std::endl;
 
-        std::cout << "fc3: (";
-        std::cout << fc3.get()->weight.size(0) << ", ";
-        std::cout << fc3.get()->weight.size(1) << ")" << std::endl;
+            std::cout << "fc3: (";
+            std::cout << fc3.get()->weight.size(0) << ", ";
+            std::cout << fc3.get()->weight.size(1) << ")" << std::endl;
 
-        std::cout << "fcSpawn: (";
-        std::cout << fcSpawn.get()->weight.size(0) << ", ";
-        std::cout << fcSpawn.get()->weight.size(1) << ")" << std::endl;
+            std::cout << "fcSpawn: (";
+            std::cout << fcSpawn.get()->weight.size(0) << ", ";
+            std::cout << fcSpawn.get()->weight.size(1) << ")" << std::endl;
+        }
     }
 
     ModelOutput forward_spawn(torch::Tensor x, torch::Tensor selected_action) {
